# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AutoSSL 证书管理API文档",
        "title": "AutoSSL API",
        "termsOfService": "http://example.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/apiv1",
    "paths": {
...
        "/auth/captcha": {
            "get": {
                "description": "生成图片验证码用于登录/注册验证",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "生成验证码",
                "responses": {
                    "200": {
                        "description": "验证码响应（包含captcha_id和image）",
                        "schema": {
                            "$ref": "#/definitions/auth.CaptchaResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "通过邮箱、密码和验证码登录系统获取访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求参数",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功响应（包含access_token和refresh_token）",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "无效请求格式",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "验证码错误/用户不存在/密码错误",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "使用刷新令牌获取新的访问令牌和刷新令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "刷新令牌",
                "parameters": [
                    {
                        "description": "刷新令牌请求参数",
                        "name": "refresh",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新成功响应（包含新的access_token和refresh_token）",
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshResponse"
                        }
                    },
                    "400": {
                        "description": "无效请求格式",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "无效刷新令牌",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "通过邮箱、密码和验证码注册新用户（第一个用户自动为管理员）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册请求参数",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "注册成功响应（包含用户ID和邮箱）",
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "无效请求格式/验证码错误",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "用户已存在",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "auth.CaptchaResponse": {
            "type": "object",
            "properties": {
                "captcha_id": {
                    "type": "string"
                },
                "error": {},
                "image": {}
            }
        },
        "auth.LoginRequest": {
            "description": "登录接口所需的请求参数",
            "type": "object",
            "required": [
                "answer",
                "captcha_id",
                "email",
                "password"
            ],
            "properties": {
                "answer": {
                    "type": "string"
                },
                "captcha_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "auth.LoginResponse": {
            "description": "登录接口返回的响应数据",
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "auth.RefreshRequest": {
            "description": "刷新令牌接口所需的请求参数",
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "auth.RefreshResponse": {
            "description": "刷新令牌接口返回的响应数据",
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "auth.RegisterRequest": {
            "description": "注册接口所需的请求参数",
            "type": "object",
            "required": [
                "answer",
                "captcha_id",
                "email",
                "password"
            ],
            "properties": {
                "answer": {
                    "type": "string"
                },
                "captcha_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "auth.RegisterResponse": {
            "description": "注册接口返回的响应数据",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}